ANDROID: Introduce compiler-side VLA detection
Variable Length Array (VLA) allows for array lengths to be determined
at run times rather than at compile times.
VLAs are covenient and a part of C99 standards but can have unintended
consequences.

Since Linux 4.20-rc1 [1], kernels are totally VLA-free. Backporting the patches
is effective still there are some hideous blocks of codes which aren't
patched by the upstream yet.

The target of this patch is to introduce a compiler side VLA detection
which will catch VLAs and will either warn about it or it will stop the
compilation. The developer can either fix the block of code by himself
or he can just look for the fix in upstream.

[1] https://www.phoronix.com/scan.php?page=news_item&px=Linux-Kills-The-VLA#:~:text=The%20Linux%20Kernel%20Is%20Now,Overhead%20%26%20Better%20For%20Clang%20%2D%20Phoronix&text=With%20the%20in%2Ddevelopment%20Linux,but%20can%20have%20unintended%20consequences.

Change-Id: I5c683dfe3d6b7427a65fbb47d25f723212545e96
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: danascape <danascape@gmail.com>
Signed-off-by: fajar <santuyz321@gmail.com>